Build edit function
Adds functionality to the edit() js function so that it can transmit
changes to the server and replace the table rows with the newer
versions. The backend updateStaging() function was modified to return a
word value rather than just a boolean, and filled with a temporary test
constant return value.